I learned that all QObjects have the method startTimer(interval) which creates a timer that creates a timerEvent every interval ms.
So I am now using that instead of a QTimer or QBasicTimer.
I also changed the way I respond to timers because I now have the timer
trigger every second, and at each event, I check to see if the session
is over or if it is time for a reminder.
If it is neither, I update a QProgressBar which displays how much time
is left as a percentage.