Make a custom widget in timerprogress.py
It's called a ProgressRing. You set a length of time in seconds and the
timer widget will display the time in HH:MM:SS counting down to 0.
There is a ring around the outside that changes color as the timer
aprroaches 0, kind of like a progress bar.

It is not complete yet but works.