In order to     1) keep track of the length of the next interval to run     2) whether the interval that triggered the finished signal is a     break or a work timer I tried making an intervals attribute in standup.py that is a list. This list contains integers which are to be send to ProgressRing as durations.
I thought this was a decent solution until I remembered that I have an
option to make a session run until the user stops it.
This would also be simple if not for breaks.

Since work and break intervals are supposed to alternate until the
session ends, it is easy enough to create a list with a set session
duration.

With no session duration, there could be cases where you need to make a
really long list.

So I'm making this change in it's own branch, I might not use it.