Refactored standup.py to make it easier to add new types of reminders
I took all of the ReminderOptions and Reminder classes out of standup
and put them in their own module, reminders.py

reminders.py also has a tuple called reminder_types which contains one
object for each ReminderOptions class.

This tuple is used in standup.py to programatically load all of the
reminder options into a QStackedWidget and a QComboBox.

I had to make some changes to the ReminderOptions class.
  - Now ReminderOptions have a name attribute. This attribute is used to
  fill the QComboBox with an appropriate string.
  - All ReminderOptions take a QLabel as an argument
  - All ReminderOptions have a method set_reminder_text that sets the
  text on the QLabel argument.