Implement StandUpWindow.transition_screen
StandUpWindow now has an additional screen called transition_screen
which is displayed after a focus interval ends. It displays the message
for the selected reminder and has a button that starts the break
interval when pressed.

I'm still thinking about how to design the UX for the transitions
between focus and break intervals so this is a prototype.