Refactor add / remove interval
Focus and break interval containers now have + and - buttons to add /
remove intervals from the end of the containers.

I chose to do this because it is easier and I don't think it will be the
end of the world to always remove the interval from the bottom of the
layout instead of letting the user select the one they want to delete.
Maybe it is though. If it's annoying I'll reconsider.